finished

+++ backend/Routes/recipes.js
@@ -9,7 +9,9 @@ const router = express.Router();
 router.post("", checkAuth,( req, res, next) => {
     const recipe = new Recipe({
         title: req.body.title,
-        method: req.body.method
+        method: req.body.method,
+        //authorisation
+        creator: req.userData.userId
     });
     recipe.save().then(createRecipe => {
         res.status(201).json({
@@ -34,14 +36,19 @@ router.get("/:id", (req, res, next) => {
 
 //Capital R for schema
 //Update a recipe.
-router.put("/:id" , (req, res, next) => {
+router.put("/:id" , checkAuth, (req, res, next) => {
     const recipe = new Recipe({
         _id: req.body.id,
         title: req.body.title,
-        method: req.body.method
-    });
-        Recipe.updateOne({_id: req.params.id}, recipe).then(result => {
-            res.status(200).json({message: "Recipe update successful"});
+        method: req.body.method,
+        creator: req.userData.userId
+    });  
+        Recipe.updateOne({_id: req.params.id, creator: req.userData.userId}, recipe).then(result => {
+            if (result.nModified > 0) {
+                res.status(200).json({ message: "Update successful!" });
+              } else {
+                res.status(401).json({ message: "Not authorized!" });
+              }
         });
     });
     
@@ -57,9 +64,12 @@ router.get("", (req, res, next) => {
 
 //send a request to post passing the id of the object you want to delete
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Recipe.deleteOne({_id: req.params.id}).then(result =>{
-        console.log(result);
-        res.status(200).json({message: 'Recipe Deleted'})
+    Recipe.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result =>{
+        if (result.n > 0) {
+            res.status(200).json({ message: "Delete successful!" });
+          } else {
+            res.status(401).json({ message: "Not authorized to delete!" });
+          }
     });
 });
